Add configuration for creating the Kenya Pool
Creates the pool from all the demographic responses from projects where we asked people if they wanted to opt-in to the pool.

I've also added an optional description field to the configuration here as I think it will be useful to be able to briefly describe what a configuration does.