Handle messages with dataset in previous_datasets
We saw this occur because a previous version of the coda sync allowed messages to be ws corrected to the dataset they are currently in. That issue was fixed in https://github.com/AfricasVoices/Test-Pipeline-Engagement-DB/pull/173, but this ensures we handle any legacy messages impacted by this and defend against similar incidents in future.

This case is problematic because it breaks the current caches. For example, if a message is in "age", then moves to "location", then to "location" again (.dataset = "location", .previous_datasets = ["age", "location"],), then the message would be silently and erroneously filtered out from analysis